Split integration TestCase into smaller classes
